Update ACL API to support ARL use cases (#34537)
* Update ACL API to support ARL use cases

ARL needs to know what attribute/command/event is being accessed
and with which interaction model action. This change brings this
knowledge into the AccessControl classes Check method for further
enhancement with ARL feature on top.

* Addressed review comments

* Addressed review comments

* Clarified RequestType enum values and purpose

* Renamed CanAccess to CanAccessEvent

* Restyled by whitespace

* Restyled by clang-format

* Fixed some build issues

- missed a spot to populate requestType and entityId
- fix unused function definition for some cases

* Fixed icd-management-server for ARL

* fixed typo in previous fix :-(

* Restyled by clang-format

* Moved ARL related checks behind ARL feature flag

Instead of validating requestType in AccessControl::Check, we will
validate it is not unknown only if the ARL feature is enabled.

* Add conditional ARL related validation

Instead of removing ARL related validation from AccessControl::Check
entirely, perform it only if the ARL feature is enabled.  As of this
commit, it is not enabled.

* Restyled by clang-format

---------

Co-authored-by: Restyled.io <commits@restyled.io>