Remove _invokeCommandWithEndpointID implementation from MTRDevice. (#35208)
This is implemented (differently) by the different subclasses.

Once this implementation is removed, the following become unused and can be
removed:

* setExpectedValues
* removeExpectedValuesForAttributePaths

Once those are removed, the following become unused and can be removed:

* _getAttributesToReportWithNewExpectedValues
* _removeExpectedValueForAttributePath

Once those are removed, the following become unused and can be removed:

* _setExpectedValue
* expectedValueNextID

At this point (since setExpectedValues was removed), the only reference to
_checkExpiredExpectedValues is in _performScheduledExpirationCheck, which is
only called from _checkExpiredExpectedValues.  So these are both unreachable and
can be removed.  Then the following become unused and can be removed:

* expirationCheckScheduled
* _reportAttributes
* _attributeDataValue
* expectedValueCache

And then _filteredAttributes and _filteredEvents are unused and can be removed.

Then both the event and attribute versions of _interestedPaths: are unused and
can be removed.

At this point, the following become unused and can be removed:

* MTRDeviceExpectedValueFieldIndex
* MTRDeviceReadRequestFieldIndex
* MTRDeviceWriteRequestFieldIndex
* MTRDeviceWorkItemBatchingID
* MTRDeviceWorkItemDuplicateTypeID

------------------------------------------------------------