TC-IDM-XX: try both PASE and CASE for basic comp tests. (#35109)
* TC-IDM-XX: try both PASE and CASE for basic comp tests.

These tests are disruptive at cert because they could run over
PASE or CASE and defaulted to PASE. This meant that they couldn't
be run in sequence with other tests that require commissioning.
Because the PASE setup required a qr or manual code, it also
meant that these needed special parameters to run.

Now:
- can use discriminator and passcode for PASE connection
- device tries both PASE and CASE and runs over whichever works
  first.

* fix CI

* Restyled by whitespace

* Restyled by clang-format

* Fix tests for NO DUT (unit and mock)

* Update src/python_testing/basic_composition_support.py

Co-authored-by: Tennessee Carmel-Veilleux <tennessee.carmelveilleux@gmail.com>

* Null terminate and add outbuf size

* Restyled by clang-format

* typo

---------

Co-authored-by: Restyled.io <commits@restyled.io>
Co-authored-by: Tennessee Carmel-Veilleux <tennessee.carmelveilleux@gmail.com>