Add matter_enable_recommended meta-setting (#34942)
* Add matter_enable_recommended meta-setting

This setting defaults to true which leaves current behavior unchanged.

However it can be set to false to achieve a more conservative / minimal set of
defaults, without having to manually disable an ever-increasing set of
features. Consider using this setting as the default value for other settings
that increase code size or add debugging / tracing or similar features that are
not always desired. Especially settings that default to true on Linux and Mac
("because you're probably building for development") should likely use this
option for their default, e.g.

matter_foo = matter_enable_recommended && (current_os == "linux" || current_os == "mac")

* Add more documentation and turn off the setting for minimal CI builds

* Move matter_enable_recommended into /config/recommended.gni

This allows it to be uesd from args.gni (i.e. in the context of a default_args
scope, where variables like current_os are not defined) as well as within a
build / toolchain context.