[ICD] Implement storing of persistent keys in PSA for ICD server (#34925)
* Refactor Key Attributes classes

Implement `keyAttributesBase` and add
persistent lifetime functionality.

* Implement `FindFreeKeySlotInRange` for PSA key slots

* Fix ICD CIP related build dependencies

* Remake `ICDMonitoringTable::Find()` not to overwrite all entry fields

If entry is not found its key handle field must not be filled with
last checked entry when using PSA as it will cause key slot to be
cleared by accident.

* Add ICD CIP and DAC key slots for PSA

* Implement `PersistICDKey` API

* Implement setting key persitence for ICD server

* Address PR comments

Modify implementation of `PersistICDKey` not to depend on type
of ICD key handle (AES, HMAC)