[nrf toup][scripts] Fix running generate for Windows
If the host system is Windows, we need to add `python3` to the
subprocess command to avoid issues like:
`OSError: [WinError 193] %1 is not a valid Win32 application`

Also, inform Windows users what they can do if paths to the
files are too long, and how to resolve issues like:
`FileNotFoundError: [WinError 206] The filename or extension is
too long`

Fixes in zap_regen_all.py for Windows
platform. The issues occur especially on Windows 11.